add sel4cp-compatible prelude header
Previously, compiling an elf for the seL4 Core Platform required modifying
the headers emitted by the Austral compiler, which assume that the program
would run as a process on a normal POSIX system.

We add a prelude template that actually works with the seL4 Core Platform,
which saves a lot of manual work. It might eventually be used by forthcoming
automated tools such as mantletool-build.

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>