implement local state
Previously, the user-defined functions notified, protected, init did not
allow the user to thread local state through.

This was very inconvenient: since Austral does not have global variables,
the only way a protection domain could maintain any persistent state was by
mapping in a memory region.

This limitation seemed necessary for IRQ handling: with a user-defined local
state type `T`, one could set `T` to a record that contains a `ChXXIrq` field,
and stash the `ChXXIrq` object without acknowledging it.

In this update we give new type signatures to notified, protected and init,
allowing them to thread through local state: they each get a state argument,
and return a modified state.

Since `protected` needs to return a `MessageInfo`, we implemented a new type
`Message[LocalState]`, which combines a MessageInfo structure with the local
state type.

For now, any local state has to be `Free`, which prevents the IRQ issue
definitively. Howver, if an argument can be made about the soundness of
the Core Platform's IRQ-related behavior in the presence of a `Linear` local
state type, this may be subject to change.

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>